Move fakes out of file
