+++ lib/features/todo/todo_repository.dart
@@ -1,15 +1,16 @@
 import 'dart:async';
 
+import 'package:todo/fake_data_source.dart';
 import 'package:todo/features/todo/todo.dart';
 import 'package:todo/features/todo/todo_entity.dart';
 import 'package:uuid/uuid.dart';
 
 // Provide a uniform way for services and view models to interact with your data
 class TodoRepository {
-  TodoRepository({required FakeLocalDataSource fakeLocalDataSource})
-      : _fakeLocalDataSource = fakeLocalDataSource;
+  TodoRepository({required FakeDataSource fakeDataSource})
+      : _fakeDataSource = fakeDataSource;
 
-  final FakeLocalDataSource _fakeLocalDataSource;
+  final FakeDataSource _fakeDataSource;
   final BehaviorSubject<List<Todo>> _todosController = BehaviorSubject([]);
 
   Stream<List<Todo>> watch() {
@@ -21,19 +22,19 @@ class TodoRepository {
     final id = uuid.v4();
     final todo = TodoEntity(id: id, title: title, completed: false);
 
-    _fakeLocalDataSource.add();
+    _fakeDataSource.add();
     _todosController.add([..._todosController.value, todo.toTodo()]);
   }
 
   Future<void> removeTodo(Todo todo) async {
-    _fakeLocalDataSource.remove();
+    _fakeDataSource.remove();
 
     _todosController
         .add(_todosController.value.where((t) => t.id != todo.id).toList());
   }
 
   Future<void> toggleDone(Todo todo) async {
-    _fakeLocalDataSource.update();
+    _fakeDataSource.update();
 
     final updatedTodos = _todosController.value.map((t) {
       if (t.id == todo.id) {
@@ -43,49 +44,5 @@ class TodoRepository {
     }).toList();
 
     _todosController.add(updatedTodos);
-  }
-}
-
-class FakeLocalDataSource {
-  Future<void> add() async {}
-
-  Future<void> remove() async {}
-
-  Future<void> update() async {}
-}
-
-class BehaviorSubject<T> {
-  // StreamController with broadcast mode
-  final StreamController<T> _controller;
-  T _currentValue;
-
-  // Constructor to initialize with an initial value
-  BehaviorSubject(T initialValue)
-      : _currentValue = initialValue,
-        _controller = StreamController<T>.broadcast();
-
-  // Getter for the current value
-  T get value => _currentValue;
-
-  // Adds a new value and broadcasts it to all listeners
-  void add(T newValue) {
-    _currentValue = newValue;
-    _controller.add(newValue);
-  }
-
-  // Exposes the stream to listen to changes
-  Stream<T> get stream => _controller.stream;
-
-  // Subscribes a listener and immediately emits the current value
-  StreamSubscription<T> listen(void Function(T) onData) {
-    final subscription = _controller.stream.listen(onData);
-    // Emit the current value immediately
-    onData(_currentValue);
-    return subscription;
-  }
-
-  // Closes the StreamController when done
-  Future<void> close() async {
-    await _controller.close();
   }
 }
